Template literals (Template strings) detection
Imagine a bit of JavaScript (JS) code like here:

let html = `
  <div class="image-box">
    <img class="image"
         src="${imageUrl}"/>
  </div>`;
// (...)
main.innerHTML = html;

You first will notice the variable called html, which constructs a bit of HTML using a Javascript template string. It also features an inclusion of another variable – imageUrl – to be used in the src attribute. The full html string is then assigned to main.innerHTML.

If we assume that the imageUrl variable is controlled by an attacker – then they might easily break out of the src attribute syntax and enter arbitrary HTML of their choosing to launch an XSS attack.

This example demonstrates how easy it is to accidentally implement a DOM XSS vulnerability: The application was expecting an image URL, but also accepts all sorts of strings, which are then parsed into HTML and JavaScripts. This is enough to enable XSS attacks.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Template_literals